Sonar: refactoring SonarResource to rest package and update API
